refactor(unauth_resource_section): cerate AddEvaluationAuthWidgets and drop AddEvaluationNoAuth
